Add handling for permalinks
Since we're using a SPA if you got there from a direct link to a post it would fail if you didn't load the angular app yet.  Using the 404 page we redirect to index.html to pick up the angular app and then let the app try to get us where they requested.

Additionally added handling of non-existent posts and some routing to land everything else we don't know about back at the post list.  Though we can add an in-app 404 later.